perf(keeper): Pre-size memory allocations with capacity hints (P3-PERF-3)
Optimize performance by pre-sizing slice allocations with appropriate
capacity hints across Oracle, DEX, and Compute keeper modules.

Changes:
- Oracle keeper (aggregation.go, abci.go, query_server.go, price.go):
  * Pre-size outlier detection results and valid prices slices
  * Pre-size snapshot collections with maxSnapshots capacity
  * Pre-size pagination results with limit capacity
  * Pre-size validator price collections with estimated validator count
  * Pre-size keys-to-delete slices for cleanup operations

- DEX keeper (pool.go, query_server.go, limit_orders.go):
  * Pre-size pool collections with MaxIterationLimit
  * Pre-size paginated query results with limit capacity
  * Pre-size order collections by owner/pool with estimated capacity
  * Pre-size expired order lists for cleanup

- Compute keeper (query_server.go):
  * Pre-size all paginated query results (providers, requests, disputes, appeals)
  * Use sanitized pagination limit for accurate capacity

Performance Impact:
- Reduces memory reallocations during append operations
- Improves throughput for high-frequency query operations
- Better memory locality and reduced GC pressure
- Particularly beneficial for aggregation and pagination paths

All DEX tests pass. Oracle and Compute modules have pre-existing test
failures unrelated to these changes.

+++ x/dex/keeper/query_server.go
@@ -79,14 +79,14 @@ func (qs queryServer) Pools(goCtx context.Context, req *types.QueryPoolsRequest)
 	}
 
 	// P3-PERF-3: Pre-size with pagination limit capacity
-	limit := defaultPaginationLimit
+	limit := uint64(defaultPaginationLimit)
 	if req.Pagination != nil && req.Pagination.Limit > 0 {
 		limit = req.Pagination.Limit
 		if limit > maxPaginationLimit {
 			limit = maxPaginationLimit
 		}
 	}
-	pools := make([]types.Pool, 0, limit)
+	pools := make([]types.Pool, 0, int(limit))
 	store := qs.Keeper.getStore(goCtx)
 	poolStore := prefix.NewStore(store, PoolKeyPrefix)
 
@@ -199,14 +199,14 @@ func (qs queryServer) LimitOrders(goCtx context.Context, req *types.QueryLimitOr
 	}
 
 	// P3-PERF-3: Pre-size with pagination limit capacity
-	limit := defaultPaginationLimit
+	limit := uint64(defaultPaginationLimit)
 	if req.Pagination != nil && req.Pagination.Limit > 0 {
 		limit = req.Pagination.Limit
 		if limit > maxPaginationLimit {
 			limit = maxPaginationLimit
 		}
 	}
-	orders := make([]types.LimitOrder, 0, limit)
+	orders := make([]types.LimitOrder, 0, int(limit))
 	store := qs.Keeper.getStore(goCtx)
 	orderStore := prefix.NewStore(store, LimitOrderKeyPrefix)
